This should test and run the validator (with no check on output yet) on the non-map StringDB examples. Unfortunately, it's getting stuck on the second test that reads from the same file.
