+++ examples/string/protein-links-detailed.py
@@ -2,6 +2,7 @@ import re
 import uuid
 
 from koza.model.biolink.model import Protein, PairwiseGeneToGeneInteraction, Predicate
+from koza.manager.data_collector import collect
 from koza.koza_runner import get_koza_app
 
 koza = get_koza_app()
@@ -22,4 +23,4 @@ for row in koza.file_registry[source_name]:
         relation='RO:0002436'
     )
 
-    koza.write(source_name, [protein_a, protein_b, pairwise_gene_to_gene_interaction])
+    collect(source_name, protein_a, protein_b, pairwise_gene_to_gene_interaction)
